+++ core/category.go
@@ -1,7 +1,6 @@
 package core
 
 import (
-	"errors"
 	"github.com/louisevanderlith/husk/hsk"
 	"github.com/louisevanderlith/husk/keys"
 	"github.com/louisevanderlith/husk/validation"
@@ -9,26 +8,16 @@ import (
 )
 
 type Category struct {
-	Name         string
-	Text         string
-	Description  string
-	PageURL      string
-	BaseCategory categories.Enum
+	Name         string          `hsk:"size(128)"`
+	Text         string          `hsk:"size(256)"`
+	Description  string          `hsk:"size(512)"`
+	PageURL      string          `hsk:"size(128)"`
+	BaseCategory categories.Enum `hsk:"null"`
 	ClientID     string
 	ImageKey     *keys.TimeKey
-	Items        []StockItem
+	OwnerKey     hsk.Key
 }
 
 func (c Category) Valid() error {
 	return validation.Struct(c)
-}
-
-func (c Category) GetItem(itemKey hsk.Key) (StockItem, int, error) {
-	for i, itm := range c.Items {
-		if itm.ItemKey == itemKey {
-			return itm, i, nil
-		}
-	}
-
-	return StockItem{}, -1, errors.New("no such item")
 }